+++ flutter_loggy_dio/lib/flutter_loggy_dio/custom_loggers.dart
@@ -2,5 +2,5 @@ import 'package:loggy/loggy.dart';
 
 mixin DioLoggy implements LoggyType {
   @override
-  Loggy<DioLoggy> get logger => Loggy<DioLoggy>('DioLoggy');
+  Loggy<DioLoggy> get loggy => Loggy<DioLoggy>('DioLoggy');
 }