+++ loggy/lib/src/types/types.dart
@@ -2,12 +2,12 @@ part of loggy;
 
 mixin NetworkLoggy implements LoggyType {
   @override
-  Loggy<NetworkLoggy> get logger =>
+  Loggy<NetworkLoggy> get loggy =>
       Loggy<NetworkLoggy>('Network Loggy - ${runtimeType.toString()}');
 }
 
 mixin UiLoggy implements LoggyType {
   @override
-  Loggy<UiLoggy> get logger =>
+  Loggy<UiLoggy> get loggy =>
       Loggy<UiLoggy>('UI Loggy - ${runtimeType.toString()}');
 }