+++ flutter_loggy/CHANGELOG.md
@@ -1,3 +1,3 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0] - 07.11.2020.
 
-* TODO: Describe initial release.
+- Initial version of Dio extensions for Loggy