+++ loggy/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 1.0.0
+## [0.1.0] - 07.11.2020.
 
-- Initial version, created by Stagehand
+- Initial version of Loggy, mixin logger for dart